% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtr_R6.R
\name{BaseContainer}
\alias{BaseContainer}
\title{BaseContainer Class}
\description{
BaseContainer class is an abstract class that is inherited by 
the Container class and ConstContainer class.
}
\examples{

## ------------------------------------------------
## Method `BaseContainer$new`
## ------------------------------------------------

ConstContainer$new()
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{data}}{is a named list containing all symbol data}

\item{\code{systemDirectory}}{is the path to GAMS System directory}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{BaseContainer$new()}}
\item \href{#method-listSymbols}{\code{BaseContainer$listSymbols()}}
\item \href{#method-listSets}{\code{BaseContainer$listSets()}}
\item \href{#method-listParameters}{\code{BaseContainer$listParameters()}}
\item \href{#method-listAliases}{\code{BaseContainer$listAliases()}}
\item \href{#method-listVariables}{\code{BaseContainer$listVariables()}}
\item \href{#method-listEquations}{\code{BaseContainer$listEquations()}}
\item \href{#method-describeSets}{\code{BaseContainer$describeSets()}}
\item \href{#method-describeAliases}{\code{BaseContainer$describeAliases()}}
\item \href{#method-describeParameters}{\code{BaseContainer$describeParameters()}}
\item \href{#method-describeVariables}{\code{BaseContainer$describeVariables()}}
\item \href{#method-describeEquations}{\code{BaseContainer$describeEquations()}}
\item \href{#method-clone}{\code{BaseContainer$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a new ConstContainer simply by initializing an object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BaseContainer$new(systemDirectory = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{systemDirectory}}{optional argument for the absolute path to 
GAMS system directory}
}
\if{html}{\out{</div>}}
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{ConstContainer$new()
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-listSymbols"></a>}}
\if{latex}{\out{\hypertarget{method-listSymbols}{}}}
\subsection{Method \code{listSymbols()}}{
list all symbols in the container if isValid = NULL
list all valid symbols in the container if  isValid = TRUE
list all invalid symbols in the container if isValid = FALSE
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BaseContainer$listSymbols(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{isValid}}{an optional logical argument}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
a vector of symbols
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-listSets"></a>}}
\if{latex}{\out{\hypertarget{method-listSets}{}}}
\subsection{Method \code{listSets()}}{
list all sets in the container if isValid = NULL
list all valid sets in the container if  isValid = TRUE
list all invalid sets in the container if isValid = FALSE
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BaseContainer$listSets(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{isValid}}{an optional logical argument}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
a vector of symbols
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-listParameters"></a>}}
\if{latex}{\out{\hypertarget{method-listParameters}{}}}
\subsection{Method \code{listParameters()}}{
list all parameters in the container if isValid = NULL
list all valid parameters in the container if  isValid = TRUE
list all invalid parameters in the container if isValid = FALSE
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BaseContainer$listParameters(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{isValid}}{an optional logical argument}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
a vector of symbols
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-listAliases"></a>}}
\if{latex}{\out{\hypertarget{method-listAliases}{}}}
\subsection{Method \code{listAliases()}}{
list all aliases in the container if isValid = NULL
list all valid aliases in the container if  isValid = TRUE
list all invalid aliases in the container if isValid = FALSE
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BaseContainer$listAliases(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{isValid}}{an optional logical argument}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
a vector of symbols
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-listVariables"></a>}}
\if{latex}{\out{\hypertarget{method-listVariables}{}}}
\subsection{Method \code{listVariables()}}{
list all variables in the container if isValid = NULL
list all valid variables in the container if  isValid = TRUE
list all invalid variables in the container if isValid = FALSE
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BaseContainer$listVariables(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{isValid}}{an optional logical argument}

\item{\code{types}}{an optional logical argument to list a subset of 
equation types}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
a vector of symbols
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-listEquations"></a>}}
\if{latex}{\out{\hypertarget{method-listEquations}{}}}
\subsection{Method \code{listEquations()}}{
list all equations in the container if isValid = NULL
list all valid equations in the container if  isValid = TRUE
list all invalid equations in the container if isValid = FALSE
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BaseContainer$listEquations(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{isValid}}{an optional logical argument}

\item{\code{types}}{an optional logical argument to list a subset of 
equation types}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
a vector of symbols
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-describeSets"></a>}}
\if{latex}{\out{\hypertarget{method-describeSets}{}}}
\subsection{Method \code{describeSets()}}{
create a summary table with descriptive statistics for Sets
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BaseContainer$describeSets(symbols = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{symbols}}{an optional argument of type string or a list of sets 
to describe. The default value is NULL which assumes all sets.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-describeAliases"></a>}}
\if{latex}{\out{\hypertarget{method-describeAliases}{}}}
\subsection{Method \code{describeAliases()}}{
create a summary table with descriptive 
statistics for Aliases
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BaseContainer$describeAliases(symbols = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{symbols}}{an optional argument of type string or a list of aliases 
to describe. The default value is NULL which assumes all aliases.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-describeParameters"></a>}}
\if{latex}{\out{\hypertarget{method-describeParameters}{}}}
\subsection{Method \code{describeParameters()}}{
create a summary table with descriptive statistics 
for Parameters
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BaseContainer$describeParameters(symbols = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{symbols}}{an optional argument of type string or a 
list of parameters to describe. The default value is 
NULL which assumes all parameters.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-describeVariables"></a>}}
\if{latex}{\out{\hypertarget{method-describeVariables}{}}}
\subsection{Method \code{describeVariables()}}{
create a summary table with descriptive 
statistics for Variables
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BaseContainer$describeVariables(symbols = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{symbols}}{an optional argument of type string 
or a list of Variables to describe. The default value 
is NULL which assumes all variables.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-describeEquations"></a>}}
\if{latex}{\out{\hypertarget{method-describeEquations}{}}}
\subsection{Method \code{describeEquations()}}{
create a summary table with descriptive 
statistics for Equations
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BaseContainer$describeEquations(symbols = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{symbols}}{an optional argument of type string or 
a list of equations to describe. The default value is 
NULL which assumes all equations.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BaseContainer$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
