% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtr_R6.R
\name{ConstContainer}
\alias{ConstContainer}
\title{ConstContainer Class}
\description{
ConstContainer class is a data-focused read-only object that 
will provide a snapshot of the data target being read. The ConstContainer 
can be created by reading a GDX file. This class is specially useful for 
the users who are only interested in post-processing data from a GAMS model 
run. The ConstContainer class inherits from an abstract
BaseContainer class. To access the functions common to Container and
ConstContainer, please use help(BaseContainer).
}
\examples{

## ------------------------------------------------
## Method `ConstContainer$new`
## ------------------------------------------------

ConstContainer$new()
}
\section{Super class}{
\code{\link[gamstransfer:BaseContainer]{gamstransfer::BaseContainer}} -> \code{ConstContainer}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{data}}{is a named list containing all symbol data}

\item{\code{systemDirectory}}{is the path to GAMS System directory}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{data}}{is a named list containing all symbol data}

\item{\code{systemDirectory}}{is the path to GAMS System directory}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{ConstContainer$new()}}
\item \href{#method-read}{\code{ConstContainer$read()}}
\item \href{#method-clone}{\code{ConstContainer$clone()}}
}
}
\if{html}{
\out{<details ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="gamstransfer" data-topic="BaseContainer" data-id="describeAliases">}\href{../../gamstransfer/html/BaseContainer.html#method-describeAliases}{\code{gamstransfer::BaseContainer$describeAliases()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="gamstransfer" data-topic="BaseContainer" data-id="describeEquations">}\href{../../gamstransfer/html/BaseContainer.html#method-describeEquations}{\code{gamstransfer::BaseContainer$describeEquations()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="gamstransfer" data-topic="BaseContainer" data-id="describeParameters">}\href{../../gamstransfer/html/BaseContainer.html#method-describeParameters}{\code{gamstransfer::BaseContainer$describeParameters()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="gamstransfer" data-topic="BaseContainer" data-id="describeSets">}\href{../../gamstransfer/html/BaseContainer.html#method-describeSets}{\code{gamstransfer::BaseContainer$describeSets()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="gamstransfer" data-topic="BaseContainer" data-id="describeVariables">}\href{../../gamstransfer/html/BaseContainer.html#method-describeVariables}{\code{gamstransfer::BaseContainer$describeVariables()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="gamstransfer" data-topic="BaseContainer" data-id="listAliases">}\href{../../gamstransfer/html/BaseContainer.html#method-listAliases}{\code{gamstransfer::BaseContainer$listAliases()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="gamstransfer" data-topic="BaseContainer" data-id="listEquations">}\href{../../gamstransfer/html/BaseContainer.html#method-listEquations}{\code{gamstransfer::BaseContainer$listEquations()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="gamstransfer" data-topic="BaseContainer" data-id="listParameters">}\href{../../gamstransfer/html/BaseContainer.html#method-listParameters}{\code{gamstransfer::BaseContainer$listParameters()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="gamstransfer" data-topic="BaseContainer" data-id="listSets">}\href{../../gamstransfer/html/BaseContainer.html#method-listSets}{\code{gamstransfer::BaseContainer$listSets()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="gamstransfer" data-topic="BaseContainer" data-id="listSymbols">}\href{../../gamstransfer/html/BaseContainer.html#method-listSymbols}{\code{gamstransfer::BaseContainer$listSymbols()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="gamstransfer" data-topic="BaseContainer" data-id="listVariables">}\href{../../gamstransfer/html/BaseContainer.html#method-listVariables}{\code{gamstransfer::BaseContainer$listVariables()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a new ConstContainer simply by initializing an object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ConstContainer$new(loadFrom = NULL, systemDirectory = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{loadFrom}}{optional argument to point to the GDX file being 
read into the Container}

\item{\code{systemDirectory}}{optional argument for the absolute path to 
GAMS system directory}
}
\if{html}{\out{</div>}}
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{ConstContainer$new()
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-read"></a>}}
\if{latex}{\out{\hypertarget{method-read}{}}}
\subsection{Method \code{read()}}{
main method to read loadFrom, can be provided 
with a list of symbols to read in subsets, `records` controls 
if symbol records are loaded or just metadata
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ConstContainer$read(loadFrom, symbols = "all", records = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{loadFrom}}{name of the file to load data from as a string}

\item{\code{symbols}}{optional argument to specify the names of the 
symbols to be read (string or a list of strings)}

\item{\code{records}}{optional logical argument to specify whether to 
read symbol records (logical)}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ConstContainer$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
