% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DomainViolation.R
\name{DomainViolation}
\alias{DomainViolation}
\title{DomainViolation Class}
\description{
Contains information about the domain violation
for a symbol.
Please visit https://www.gams.com/latest/docs/API_R_GAMSTRANSFER.html 
for detailed documentation of this package.
}
\section{Fields}{

\describe{
\item{\code{symbol}}{symbol name}

\item{\code{dimension}}{dimension in which domain violation is present}

\item{\code{domain}}{domain name}

\item{\code{violations}}{vector of violations}
}}

