% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Symbol.R
\name{.Symbol}
\alias{.Symbol}
\title{Symbol Abstract Class}
\description{
An abstract symbol class from
which the classes Set, Parameter, Variable,
and Equation are inherited.
Please visit https://transfer-r.readthedocs.io/en/latest/
for detailed documentation of this package.
}
