% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamstransfer-package.R
\docType{package}
\name{gamstransfer-package}
\alias{gamstransfer}
\alias{gamstransfer-package}
\title{gamstransfer: A Data Interface Between 'GAMS' and R}
\description{
Read, analyze, modify, and write 'GAMS' (General Algebraic Modeling System) data. The main focus of 'gamstransfer' is the highly efficient transfer of data with 'GAMS' \url{https://www.gams.com/}, while keeping these operations as simple as possible for the user. The transfer of data usually takes place via an intermediate GDX (GAMS Data Exchange) file. Additionally, 'gamstransfer' provides utility functions to get an overview of 'GAMS' data and to check its validity.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/GAMS-dev/transfer-r/tree/main/gamstransfer}
  \item Report bugs at \url{https://github.com/GAMS-dev/transfer-r/issues}
}

}
\author{
\strong{Maintainer}: Atharv Bhosekar \email{abhosekar@gams.com}

Other contributors:
\itemize{
  \item GAMS Development Corp. \email{support@gams.com} [copyright holder, funder]
  \item GAMS Software GmbH \email{support@gams.com} [copyright holder, funder]
}

}
\keyword{internal}
