% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Parameter.R
\name{Parameter}
\alias{Parameter}
\title{Parameter Class}
\description{
A class for Parameter objects. This class inherits from an abstract
Symbol class.The documentation for methods common to all symbols can be accessed
via help(.Symbol).
Please visit https://transfer-r.readthedocs.io/en/latest/
for detailed documentation of this package.
}
\examples{
# create a container
m <- Container$new()
# add a Parameter
p <- Parameter$new(m, "p")
# access records
p_recs <- p$records
}
