% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Super.R
\name{readGDX}
\alias{readGDX}
\title{readGDX}
\usage{
readGDX(loadFrom, symbols = NULL, records = TRUE)
}
\arguments{
\item{loadFrom}{name of the GDX file being read (string)}

\item{symbols}{optional argument - vector of strings
containing the symbol names to be read}

\item{records}{optional logical argument - TRUE (default) to read 
the symbol records, FALSE to only read the meta data}
}
\description{
read a GDX file to a list without creating symbol 
or container objects
}
