% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpecialValues.R
\docType{data}
\name{SpecialValues}
\alias{SpecialValues}
\title{SpecialValues list object}
\format{
An object of class \code{list} of length 10.
}
\usage{
SpecialValues
}
\description{
This list contains GAMS special values and
helper functions to check if a given value is a GAMS
special value.
Please visit https://www.gams.com/latest/docs/API_R_GAMSTRANSFER.html 
for detailed documentation of this package.
}
\examples{
# check the value of GAMS special value NA
SpecialValues[["NA"]]
# check the value of GAMS special value EPS
SpecialValues[["EPS"]]
# check if a value is GAMS special value `NA`
SpecialValues$isNA(0)
}
\keyword{datasets}
