% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UniverseAlias.R
\name{UniverseAlias}
\alias{UniverseAlias}
\title{UniverseAlias Class}
\description{
A class for Alias objects that are aliased to the Universe set.
Please visit https://www.gams.com/latest/docs/API_R_GAMSTRANSFER.html 
for detailed documentation of this package.
}
