% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Super.R
\name{writeGDX}
\alias{writeGDX}
\title{writeGDX}
\usage{
writeGDX(
  writeList,
  writeTo,
  symbols = NULL,
  compress = FALSE,
  uelPriority = NULL,
  mode = NULL
)
}
\arguments{
\item{writeList}{list containing symbol data and metadata}

\item{writeTo}{name of the output GDX file}

\item{symbols}{optional argument - vector of strings
containing the symbol names to be read}

\item{compress}{optional logical argument. TRUE to produce a 
compressed GDX file}

\item{uelPriority}{Specify the priority UELs}

\item{mode}{optional string argument to specify the write 
mode ("string", "mapped").
Please visit https://www.gams.com/latest/docs/API_R_GAMSTRANSFER.html 
for detailed documentation of this package.}
}
\description{
write a GDX file from a list containing symbol data
and metadata
}
