% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtr_R6.R
\name{Container}
\alias{Container}
\title{Container object}
\description{
Container is the unified object that contains data
}
\examples{
Container$new()
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{data}}{is a list containing all symbols}

\item{\code{systemDirectory}}{is the path to GAMS System directory}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Container$new()}}
\item \href{#method-read}{\code{Container$read()}}
\item \href{#method-getUniverseSet}{\code{Container$getUniverseSet()}}
\item \href{#method-removeSymbols}{\code{Container$removeSymbols()}}
\item \href{#method-renameSymbol}{\code{Container$renameSymbol()}}
\item \href{#method-listSymbols}{\code{Container$listSymbols()}}
\item \href{#method-listSets}{\code{Container$listSets()}}
\item \href{#method-listParameters}{\code{Container$listParameters()}}
\item \href{#method-listAliases}{\code{Container$listAliases()}}
\item \href{#method-listVariables}{\code{Container$listVariables()}}
\item \href{#method-listEquations}{\code{Container$listEquations()}}
\item \href{#method-addSet}{\code{Container$addSet()}}
\item \href{#method-addParameter}{\code{Container$addParameter()}}
\item \href{#method-addVariable}{\code{Container$addVariable()}}
\item \href{#method-addEquation}{\code{Container$addEquation()}}
\item \href{#method-addAlias}{\code{Container$addAlias()}}
\item \href{#method-describeSets}{\code{Container$describeSets()}}
\item \href{#method-describeParameters}{\code{Container$describeParameters()}}
\item \href{#method-describeVariables}{\code{Container$describeVariables()}}
\item \href{#method-describeEquations}{\code{Container$describeEquations()}}
\item \href{#method-printSpecialValues}{\code{Container$printSpecialValues()}}
\item \href{#method-write}{\code{Container$write()}}
\item \href{#method-reOrderSymbols}{\code{Container$reOrderSymbols()}}
\item \href{#method-isValid}{\code{Container$isValid()}}
\item \href{#method-.linkDomainCategories}{\code{Container$.linkDomainCategories()}}
\item \href{#method-clone}{\code{Container$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a new container
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Container$new(loadFrom = NULL, systemDirectory = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{loadFrom}}{name of the GDX file to load data from}

\item{\code{systemDirectory}}{optional argument for the path to GAMS System directory}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
read a file using
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-read"></a>}}
\if{latex}{\out{\hypertarget{method-read}{}}}
\subsection{Method \code{read()}}{
read data from a GDX file
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Container$read(loadFrom, symbols = "all", values = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{loadFrom}}{name of the file to load data from}

\item{\code{symbols}}{optional argument to specify the names of the symbols to be read}

\item{\code{values}}{optional boolean argument to specify whether to read symbol records}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
`$read()` reads a file
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getUniverseSet"></a>}}
\if{latex}{\out{\hypertarget{method-getUniverseSet}{}}}
\subsection{Method \code{getUniverseSet()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Container$getUniverseSet()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-removeSymbols"></a>}}
\if{latex}{\out{\hypertarget{method-removeSymbols}{}}}
\subsection{Method \code{removeSymbols()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Container$removeSymbols(name = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-renameSymbol"></a>}}
\if{latex}{\out{\hypertarget{method-renameSymbol}{}}}
\subsection{Method \code{renameSymbol()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Container$renameSymbol(oldName = NULL, newName = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-listSymbols"></a>}}
\if{latex}{\out{\hypertarget{method-listSymbols}{}}}
\subsection{Method \code{listSymbols()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Container$listSymbols(isValid = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-listSets"></a>}}
\if{latex}{\out{\hypertarget{method-listSets}{}}}
\subsection{Method \code{listSets()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Container$listSets(isValid = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-listParameters"></a>}}
\if{latex}{\out{\hypertarget{method-listParameters}{}}}
\subsection{Method \code{listParameters()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Container$listParameters(isValid = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-listAliases"></a>}}
\if{latex}{\out{\hypertarget{method-listAliases}{}}}
\subsection{Method \code{listAliases()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Container$listAliases(isValid = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-listVariables"></a>}}
\if{latex}{\out{\hypertarget{method-listVariables}{}}}
\subsection{Method \code{listVariables()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Container$listVariables(isValid = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-listEquations"></a>}}
\if{latex}{\out{\hypertarget{method-listEquations}{}}}
\subsection{Method \code{listEquations()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Container$listEquations(isValid = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-addSet"></a>}}
\if{latex}{\out{\hypertarget{method-addSet}{}}}
\subsection{Method \code{addSet()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Container$addSet(
  name,
  domain = "*",
  isSingleton = FALSE,
  records = NULL,
  domainForwarding = FALSE,
  description = ""
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-addParameter"></a>}}
\if{latex}{\out{\hypertarget{method-addParameter}{}}}
\subsection{Method \code{addParameter()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Container$addParameter(
  name,
  domain = list(),
  records = NULL,
  domainForwarding = FALSE,
  description = ""
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-addVariable"></a>}}
\if{latex}{\out{\hypertarget{method-addVariable}{}}}
\subsection{Method \code{addVariable()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Container$addVariable(
  name,
  type = "free",
  domain = list(),
  records = NULL,
  domainForwarding = FALSE,
  description = ""
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-addEquation"></a>}}
\if{latex}{\out{\hypertarget{method-addEquation}{}}}
\subsection{Method \code{addEquation()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Container$addEquation(
  name,
  type,
  domain = list(),
  records = NULL,
  domainForwarding = FALSE,
  description = ""
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-addAlias"></a>}}
\if{latex}{\out{\hypertarget{method-addAlias}{}}}
\subsection{Method \code{addAlias()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Container$addAlias(name, aliasWith)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-describeSets"></a>}}
\if{latex}{\out{\hypertarget{method-describeSets}{}}}
\subsection{Method \code{describeSets()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Container$describeSets(symbols = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-describeParameters"></a>}}
\if{latex}{\out{\hypertarget{method-describeParameters}{}}}
\subsection{Method \code{describeParameters()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Container$describeParameters(symbols = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-describeVariables"></a>}}
\if{latex}{\out{\hypertarget{method-describeVariables}{}}}
\subsection{Method \code{describeVariables()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Container$describeVariables(symbols = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-describeEquations"></a>}}
\if{latex}{\out{\hypertarget{method-describeEquations}{}}}
\subsection{Method \code{describeEquations()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Container$describeEquations(symbols = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-printSpecialValues"></a>}}
\if{latex}{\out{\hypertarget{method-printSpecialValues}{}}}
\subsection{Method \code{printSpecialValues()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Container$printSpecialValues()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-write"></a>}}
\if{latex}{\out{\hypertarget{method-write}{}}}
\subsection{Method \code{write()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Container$write(gdxout, compress = FALSE, uelPriority = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-reOrderSymbols"></a>}}
\if{latex}{\out{\hypertarget{method-reOrderSymbols}{}}}
\subsection{Method \code{reOrderSymbols()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Container$reOrderSymbols()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-isValid}{}}}
\subsection{Method \code{isValid()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Container$isValid(verbose = FALSE, force = FALSE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-.linkDomainCategories"></a>}}
\if{latex}{\out{\hypertarget{method-.linkDomainCategories}{}}}
\subsection{Method \code{.linkDomainCategories()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Container$.linkDomainCategories()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Container$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
