% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Container.R
\name{Container}
\alias{Container}
\title{Container Class}
\value{
a list of object references to symbols
}
\description{
The main object class within GAMS Transfer is called 
Container. The Container is the vessel that allows symbols to be 
linked together (through their domain definitions), it enables 
implicit set definitions, it enables structural manipulations of the 
data (matrix generation), and it allows the user to perform different 
read/write operations.
}
\examples{

## ------------------------------------------------
## Method `Container$new`
## ------------------------------------------------

Container$new()
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{data}}{is a named list containing all symbol data}

\item{\code{systemDirectory}}{is the path to GAMS System directory}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Container-new}{\code{Container$new()}}
\item \href{#method-Container-[}{\code{Container$[()}}
\item \href{#method-Container-[<-}{\code{Container$[<-()}}
\item \href{#method-Container-format}{\code{Container$format()}}
\item \href{#method-Container-hasSymbols}{\code{Container$hasSymbols()}}
\item \href{#method-Container-getSymbolNames}{\code{Container$getSymbolNames()}}
\item \href{#method-Container-listSymbols}{\code{Container$listSymbols()}}
\item \href{#method-Container-listSets}{\code{Container$listSets()}}
\item \href{#method-Container-listParameters}{\code{Container$listParameters()}}
\item \href{#method-Container-listAliases}{\code{Container$listAliases()}}
\item \href{#method-Container-listVariables}{\code{Container$listVariables()}}
\item \href{#method-Container-listEquations}{\code{Container$listEquations()}}
\item \href{#method-Container-describeSets}{\code{Container$describeSets()}}
\item \href{#method-Container-describeAliases}{\code{Container$describeAliases()}}
\item \href{#method-Container-describeParameters}{\code{Container$describeParameters()}}
\item \href{#method-Container-describeVariables}{\code{Container$describeVariables()}}
\item \href{#method-Container-describeEquations}{\code{Container$describeEquations()}}
\item \href{#method-Container-read}{\code{Container$read()}}
\item \href{#method-Container-getUniverseSet}{\code{Container$getUniverseSet()}}
\item \href{#method-Container-removeSymbols}{\code{Container$removeSymbols()}}
\item \href{#method-Container-renameSymbol}{\code{Container$renameSymbol()}}
\item \href{#method-Container-addSet}{\code{Container$addSet()}}
\item \href{#method-Container-addParameter}{\code{Container$addParameter()}}
\item \href{#method-Container-addVariable}{\code{Container$addVariable()}}
\item \href{#method-Container-addEquation}{\code{Container$addEquation()}}
\item \href{#method-Container-addAlias}{\code{Container$addAlias()}}
\item \href{#method-Container-addUniverseAlias}{\code{Container$addUniverseAlias()}}
\item \href{#method-Container-getSymbols}{\code{Container$getSymbols()}}
\item \href{#method-Container-write}{\code{Container$write()}}
\item \href{#method-Container-reorderSymbols}{\code{Container$reorderSymbols()}}
\item \href{#method-Container-isValid}{\code{Container$isValid()}}
\item \href{#method-Container-.linkDomainCategories}{\code{Container$.linkDomainCategories()}}
\item \href{#method-Container-getUELs}{\code{Container$getUELs()}}
\item \href{#method-Container-removeUELs}{\code{Container$removeUELs()}}
\item \href{#method-Container-renameUELs}{\code{Container$renameUELs()}}
\item \href{#method-Container-getDomainViolations}{\code{Container$getDomainViolations()}}
\item \href{#method-Container-hasDomainViolations}{\code{Container$hasDomainViolations()}}
\item \href{#method-Container-countDomainViolations}{\code{Container$countDomainViolations()}}
\item \href{#method-Container-dropDomainViolations}{\code{Container$dropDomainViolations()}}
\item \href{#method-Container-countDuplicateRecords}{\code{Container$countDuplicateRecords()}}
\item \href{#method-Container-hasDuplicateRecords}{\code{Container$hasDuplicateRecords()}}
\item \href{#method-Container-dropDuplicateRecords}{\code{Container$dropDuplicateRecords()}}
\item \href{#method-Container-equals}{\code{Container$equals()}}
\item \href{#method-Container-copy}{\code{Container$copy()}}
\item \href{#method-Container-clone}{\code{Container$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Container-new"></a>}}
\if{latex}{\out{\hypertarget{method-Container-new}{}}}
\subsection{Method \code{new()}}{
Create a new container simply by initializing an object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Container$new(loadFrom = NULL, systemDirectory = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{loadFrom}}{optional argument to point to the GDX file being 
read into the Container}

\item{\code{systemDirectory}}{optional argument for the absolute path to 
GAMS system directory}
}
\if{html}{\out{</div>}}
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{Container$new()
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Container-["></a>}}
\if{latex}{\out{\hypertarget{method-Container-[}{}}}
\subsection{Method \code{[()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Container$[(key)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Container-[<-"></a>}}
\if{latex}{\out{\hypertarget{method-Container-[<-}{}}}
\subsection{Method \code{[<-()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Container$[<-(key, values)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Container-format"></a>}}
\if{latex}{\out{\hypertarget{method-Container-format}{}}}
\subsection{Method \code{format()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Container$format(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Container-hasSymbols"></a>}}
\if{latex}{\out{\hypertarget{method-Container-hasSymbols}{}}}
\subsection{Method \code{hasSymbols()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Container$hasSymbols(names)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Container-getSymbolNames"></a>}}
\if{latex}{\out{\hypertarget{method-Container-getSymbolNames}{}}}
\subsection{Method \code{getSymbolNames()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Container$getSymbolNames(names)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Container-listSymbols"></a>}}
\if{latex}{\out{\hypertarget{method-Container-listSymbols}{}}}
\subsection{Method \code{listSymbols()}}{
list all symbols in the container if isValid = NULL
list all valid symbols in the container if  isValid = TRUE
list all invalid symbols in the container if isValid = FALSE
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Container$listSymbols(isValid = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{isValid}}{an optional logical argument}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
a vector of symbols
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Container-listSets"></a>}}
\if{latex}{\out{\hypertarget{method-Container-listSets}{}}}
\subsection{Method \code{listSets()}}{
list all sets in the container if isValid = NULL
list all valid sets in the container if  isValid = TRUE
list all invalid sets in the container if isValid = FALSE
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Container$listSets(isValid = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{isValid}}{an optional logical argument}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
a vector of symbols
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Container-listParameters"></a>}}
\if{latex}{\out{\hypertarget{method-Container-listParameters}{}}}
\subsection{Method \code{listParameters()}}{
list all parameters in the container if isValid = NULL
list all valid parameters in the container if  isValid = TRUE
list all invalid parameters in the container if isValid = FALSE
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Container$listParameters(isValid = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{isValid}}{an optional logical argument}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
a vector of symbols
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Container-listAliases"></a>}}
\if{latex}{\out{\hypertarget{method-Container-listAliases}{}}}
\subsection{Method \code{listAliases()}}{
list all aliases in the container if isValid = NULL
list all valid aliases in the container if  isValid = TRUE
list all invalid aliases in the container if isValid = FALSE
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Container$listAliases(isValid = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{isValid}}{an optional logical argument}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
a vector of symbols
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Container-listVariables"></a>}}
\if{latex}{\out{\hypertarget{method-Container-listVariables}{}}}
\subsection{Method \code{listVariables()}}{
list all variables in the container if isValid = NULL
list all valid variables in the container if  isValid = TRUE
list all invalid variables in the container if isValid = FALSE
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Container$listVariables(isValid = NULL, types = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{isValid}}{an optional logical argument}

\item{\code{types}}{an optional logical argument to list a subset of 
equation types}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
a vector of symbols
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Container-listEquations"></a>}}
\if{latex}{\out{\hypertarget{method-Container-listEquations}{}}}
\subsection{Method \code{listEquations()}}{
list all equations in the container if isValid = NULL
list all valid equations in the container if  isValid = TRUE
list all invalid equations in the container if isValid = FALSE
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Container$listEquations(isValid = NULL, types = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{isValid}}{an optional logical argument}

\item{\code{types}}{an optional logical argument to list a subset of 
equation types}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
a vector of symbols
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Container-describeSets"></a>}}
\if{latex}{\out{\hypertarget{method-Container-describeSets}{}}}
\subsection{Method \code{describeSets()}}{
create a summary table with descriptive statistics for Sets
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Container$describeSets(symbols = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{symbols}}{an optional argument of type string or a list of sets 
to describe. The default value is NULL which assumes all sets.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Container-describeAliases"></a>}}
\if{latex}{\out{\hypertarget{method-Container-describeAliases}{}}}
\subsection{Method \code{describeAliases()}}{
create a summary table with descriptive 
statistics for Aliases
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Container$describeAliases(symbols = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{symbols}}{an optional argument of type string or a list of aliases 
to describe. The default value is NULL which assumes all aliases.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Container-describeParameters"></a>}}
\if{latex}{\out{\hypertarget{method-Container-describeParameters}{}}}
\subsection{Method \code{describeParameters()}}{
create a summary table with descriptive statistics 
for Parameters
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Container$describeParameters(symbols = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{symbols}}{an optional argument of type string or a 
list of parameters to describe. The default value is 
NULL which assumes all parameters.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Container-describeVariables"></a>}}
\if{latex}{\out{\hypertarget{method-Container-describeVariables}{}}}
\subsection{Method \code{describeVariables()}}{
create a summary table with descriptive 
statistics for Variables
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Container$describeVariables(symbols = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{symbols}}{an optional argument of type string 
or a list of Variables to describe. The default value 
is NULL which assumes all variables.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Container-describeEquations"></a>}}
\if{latex}{\out{\hypertarget{method-Container-describeEquations}{}}}
\subsection{Method \code{describeEquations()}}{
create a summary table with descriptive 
statistics for Equations
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Container$describeEquations(symbols = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{symbols}}{an optional argument of type string or 
a list of equations to describe. The default value is 
NULL which assumes all equations.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Container-read"></a>}}
\if{latex}{\out{\hypertarget{method-Container-read}{}}}
\subsection{Method \code{read()}}{
main method to read loadFrom, can be provided 
with a list of symbols to read in subsets, `records` controls 
if symbol records are loaded or just metadata
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Container$read(loadFrom, symbols = NULL, records = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{loadFrom}}{name of the file to load data from as a string}

\item{\code{symbols}}{optional argument to specify the names of the 
symbols to be read (string or a list of strings)}

\item{\code{records}}{optional logical argument to specify whether to 
read symbol records (logical)}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Container-getUniverseSet"></a>}}
\if{latex}{\out{\hypertarget{method-Container-getUniverseSet}{}}}
\subsection{Method \code{getUniverseSet()}}{
provides a universe for all symbols
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Container$getUniverseSet()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Container-removeSymbols"></a>}}
\if{latex}{\out{\hypertarget{method-Container-removeSymbols}{}}}
\subsection{Method \code{removeSymbols()}}{
removes symbols from the Container
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Container$removeSymbols(symbols = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{symbols}}{a string specifying the symbol name or a list of symbols 
to be removed from the container}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Container-renameSymbol"></a>}}
\if{latex}{\out{\hypertarget{method-Container-renameSymbol}{}}}
\subsection{Method \code{renameSymbol()}}{
rename a symbol in the Container
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Container$renameSymbol(oldName = NULL, newName = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{oldName}}{string specifying the old name of the symbol}

\item{\code{newName}}{string specifying the new name of the symbol}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Container-addSet"></a>}}
\if{latex}{\out{\hypertarget{method-Container-addSet}{}}}
\subsection{Method \code{addSet()}}{
There are two different ways to create a GAMS set and 
add it to a Container. One is using the Set constructor and 
the other is using addSet method which calls the Set constructor
internally.
addSet is a Container method to add a Set.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Container$addSet(
  name,
  domain = "*",
  isSingleton = FALSE,
  records = NULL,
  domainForwarding = FALSE,
  description = ""
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{string argument for name of the set}

\item{\code{domain}}{an optional argument specifying a list of strings, 
a string. default value is "*".}

\item{\code{isSingleton}}{an optional logical argument specifying if a set
is singleton. Default value is FALSE.}

\item{\code{records}}{specify set records as a string vector or a dataframe.}

\item{\code{domainForwarding}}{an optional logical argument to specify 
domain forwarding. Default value is FALSE.}

\item{\code{description}}{string specifying description for the set}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
a Set object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Container-addParameter"></a>}}
\if{latex}{\out{\hypertarget{method-Container-addParameter}{}}}
\subsection{Method \code{addParameter()}}{
There are two different ways to create a GAMS parameter and 
add it to a Container. One is using the Parameter constructor and 
the other is using addParameter method which calls the 
Parameter constructor internally.
addParameter is a Container method to add a Parameter.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Container$addParameter(
  name,
  domain = NULL,
  records = NULL,
  domainForwarding = FALSE,
  description = ""
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{string argument for name of the Parameter}

\item{\code{domain}}{an optional argument specifying a list of strings, 
a string. default value is NULL.}

\item{\code{records}}{specify set records as a vector or a dataframe.}

\item{\code{domainForwarding}}{an optional logical argument to specify 
domain forwarding. Default value is FALSE.}

\item{\code{description}}{string specifying description for the set}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
a Parameter object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Container-addVariable"></a>}}
\if{latex}{\out{\hypertarget{method-Container-addVariable}{}}}
\subsection{Method \code{addVariable()}}{
There are two different ways to create a GAMS Variable and 
add it to a Container. One is using the Variable constructor and 
the other is using addVariable method which calls the Parameter 
constructor internally.
addVariable is a Container method to add a Variable.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Container$addVariable(
  name,
  type = "free",
  domain = NULL,
  records = NULL,
  domainForwarding = FALSE,
  description = ""
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{string argument for name of the Variable}

\item{\code{type}}{Type of variable being created [binary, integer, 
positive, negative, free, sos1, sos2, semicont, semiint]. The default
is "free"}

\item{\code{domain}}{an optional argument specifying a list of strings, 
a string. default value is NULL.}

\item{\code{records}}{specify set records as a vector or a dataframe.}

\item{\code{domainForwarding}}{an optional logical argument to specify 
domain forwarding. Default value is FALSE.}

\item{\code{description}}{string specifying description for the set}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
a Variable object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Container-addEquation"></a>}}
\if{latex}{\out{\hypertarget{method-Container-addEquation}{}}}
\subsection{Method \code{addEquation()}}{
There are two different ways to create a GAMS Equation and 
add it to a Container. One is using the Equation constructor and 
the other is using addEquation method which calls the Equation 
constructor internally.
addEquation is a Container method to add a Equation.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Container$addEquation(
  name,
  type,
  domain = NULL,
  records = NULL,
  domainForwarding = FALSE,
  description = ""
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{string argument for name of the Equation}

\item{\code{type}}{Type of equation being created [eq (or E/e), geq 
(or G/g), leq (or L/l), nonbinding (or N/n), external (or X/x)]}

\item{\code{domain}}{an optional argument specifying a list of strings, 
a string. default value is "*".}

\item{\code{records}}{specify set records as a vector or a dataframe.}

\item{\code{domainForwarding}}{an optional logical argument to specify 
domain forwarding. Default value is FALSE.}

\item{\code{description}}{string specifying description for the set}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
a Equation object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Container-addAlias"></a>}}
\if{latex}{\out{\hypertarget{method-Container-addAlias}{}}}
\subsection{Method \code{addAlias()}}{
There are two different ways to create a GAMS Alias and 
add it to a Container. One is using the Alias constructor and 
the other is using addAlias method which calls the Alias 
constructor internally.
addAlias is a Container method to add a Alias.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Container$addAlias(name, aliasWith)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{string argument for name of the Alias}

\item{\code{aliasWith}}{string argument for the set/alias we want to add
an alias for}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Container-addUniverseAlias"></a>}}
\if{latex}{\out{\hypertarget{method-Container-addUniverseAlias}{}}}
\subsection{Method \code{addUniverseAlias()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Container$addUniverseAlias(name)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Container-getSymbols"></a>}}
\if{latex}{\out{\hypertarget{method-Container-getSymbols}{}}}
\subsection{Method \code{getSymbols()}}{
returns a list of object references for `Symbols`
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Container$getSymbols(symbols)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{symbols}}{character, string, or vector of Symbols for which 
the user wants object references}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Container-write"></a>}}
\if{latex}{\out{\hypertarget{method-Container-write}{}}}
\subsection{Method \code{write()}}{
a write method to write to a writeTo GDX file
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Container$write(writeTo, symbols = NULL, compress = FALSE, uelPriority = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{writeTo}}{name of the GDX file to write to}

\item{\code{compress}}{write tge GDX file in compressed format by setting
compress = TRUE.}

\item{\code{uelPriority}}{Advanced users might want to specify an order 
to their UEL list (i.e., the universe set); The UEL 
ordering follows that dictated by the data. As a convenience, it 
is possible to prepend the UEL list with a user specified order 
using the uel_priority argument.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Container-reorderSymbols"></a>}}
\if{latex}{\out{\hypertarget{method-Container-reorderSymbols}{}}}
\subsection{Method \code{reorderSymbols()}}{
reorder symbols in order to avoid domain violations
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Container$reorderSymbols()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Container-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-Container-isValid}{}}}
\subsection{Method \code{isValid()}}{
TRUE if all the symbols is in the Container are 
valid, throw exceptions if verbose=True, check all symbols if 
force=TRUE.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Container$isValid(symbols = NULL, verbose = FALSE, force = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{verbose}}{type logical}

\item{\code{force}}{type logical}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Container-.linkDomainCategories"></a>}}
\if{latex}{\out{\hypertarget{method-Container-.linkDomainCategories}{}}}
\subsection{Method \code{.linkDomainCategories()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Container$.linkDomainCategories()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Container-getUELs"></a>}}
\if{latex}{\out{\hypertarget{method-Container-getUELs}{}}}
\subsection{Method \code{getUELs()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Container$getUELs(symbols = NULL, ignoreUnused = FALSE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Container-removeUELs"></a>}}
\if{latex}{\out{\hypertarget{method-Container-removeUELs}{}}}
\subsection{Method \code{removeUELs()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Container$removeUELs(uels = NULL, symbols = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Container-renameUELs"></a>}}
\if{latex}{\out{\hypertarget{method-Container-renameUELs}{}}}
\subsection{Method \code{renameUELs()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Container$renameUELs(uels, symbols = NULL, allowMerge = FALSE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Container-getDomainViolations"></a>}}
\if{latex}{\out{\hypertarget{method-Container-getDomainViolations}{}}}
\subsection{Method \code{getDomainViolations()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Container$getDomainViolations(symbols = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Container-hasDomainViolations"></a>}}
\if{latex}{\out{\hypertarget{method-Container-hasDomainViolations}{}}}
\subsection{Method \code{hasDomainViolations()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Container$hasDomainViolations(symbols = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Container-countDomainViolations"></a>}}
\if{latex}{\out{\hypertarget{method-Container-countDomainViolations}{}}}
\subsection{Method \code{countDomainViolations()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Container$countDomainViolations(symbols = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Container-dropDomainViolations"></a>}}
\if{latex}{\out{\hypertarget{method-Container-dropDomainViolations}{}}}
\subsection{Method \code{dropDomainViolations()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Container$dropDomainViolations(symbols = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Container-countDuplicateRecords"></a>}}
\if{latex}{\out{\hypertarget{method-Container-countDuplicateRecords}{}}}
\subsection{Method \code{countDuplicateRecords()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Container$countDuplicateRecords(symbols = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Container-hasDuplicateRecords"></a>}}
\if{latex}{\out{\hypertarget{method-Container-hasDuplicateRecords}{}}}
\subsection{Method \code{hasDuplicateRecords()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Container$hasDuplicateRecords(symbols = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Container-dropDuplicateRecords"></a>}}
\if{latex}{\out{\hypertarget{method-Container-dropDuplicateRecords}{}}}
\subsection{Method \code{dropDuplicateRecords()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Container$dropDuplicateRecords(symbols = NULL, keep = "first")}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Container-equals"></a>}}
\if{latex}{\out{\hypertarget{method-Container-equals}{}}}
\subsection{Method \code{equals()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Container$equals(other, verbose = FALSE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Container-copy"></a>}}
\if{latex}{\out{\hypertarget{method-Container-copy}{}}}
\subsection{Method \code{copy()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Container$copy(destination, symbols = NULL, overwrite = FALSE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Container-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Container-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Container$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
