% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Container.R
\name{Container}
\alias{Container}
\title{Container Class}
\arguments{
\item{loadFrom}{optional argument to point to the GDX file being 
read into the Container}
}
\description{
The main object class within GAMS Transfer is called 
Container. The Container is the vessel that allows symbols to be 
linked together (through their domain definitions), it enables 
implicit set definitions, it enables structural manipulations of the 
data (matrix generation), and it allows the user to perform different 
read/write operations.
Please visit https://www.gams.com/latest/docs/API_R_GAMSTRANSFER.html 
for detailed documentation of this package.

Create a new container simply by initializing an object.
}
\section{Fields}{

\describe{
\item{\code{data}}{is a named list containing all symbol data}

\item{\code{systemDirectory}}{is the path to GAMS System directory}
}}

