% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Equation.R
\name{Equation}
\alias{Equation}
\title{Equation Class}
\description{
A class for Equation objects. This class inherits from an abstract
symbol class.The documentation for methods common to all symbols can be
accessed via help(.Symbol).
Please visit https://transfer-r.readthedocs.io/en/latest/
for detailed documentation of this package.
}
\examples{
# create a container
m <- Container$new()
# add a Variable
e <- Equation$new(m, "v", type = "eq")
# access records
e_recs <- e$records
}
