% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Equation.R
\name{Equation}
\alias{Equation}
\title{Equation Class}
\description{
A class for Equation objects. This class inherits from an abstract 
symbol class.The documentation for methods common to all symbols can be
accessed via help(.Symbol).
Please visit https://www.gams.com/latest/docs/API_R_GAMSTRANSFER.html 
for detailed documentation of this package.
}
