% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamstransfer-package.R
\docType{package}
\name{gamstransfer-package}
\alias{gamstransfer}
\alias{gamstransfer-package}
\title{gamstransfer: A Data Interface Between 'GAMS' and R}
\description{
Read, analyze, modify, and write 'GAMS' (General Algebraic Modeling System) data using 'R'. The main focus of 'gamstransfer' is the highly efficient transfer of data between 'GAMS' \url{https://www.gams.com/} and R, while keeping those operations as simple as possible for the user. With the help of 'gamstransfer', users can add, manipulate, read or write data to 'GAMS' usually via an intermediate GDX (GAMS Data Exchange) file. 'gamstransfer' also provides utility functions to quickly and easily analyze 'GAMS' data.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/GAMS-dev/transfer-r/tree/main/gamstransfer}
  \item Report bugs at \url{https://github.com/GAMS-dev/transfer-r/issues}
}

}
\author{
\strong{Maintainer}: Atharv Bhosekar \email{abhosekar@gams.com}

Other contributors:
\itemize{
  \item GAMS Development Corp. \email{support@gams.com} [copyright holder, funder]
  \item GAMS Software GmbH \email{support@gams.com} [copyright holder, funder]
}

}
\keyword{internal}
