% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Set.R
\name{Set}
\alias{Set}
\title{Set Class}
\description{
A class for Set objects. This class inherits from an abstract Symbol class.
The documentation for methods common to all symbols can be accessed via help(.Symbol).
Please visit https://transfer-r.readthedocs.io/en/latest/
for detailed documentation of this package.
}
\examples{
# create a container
m <- Container$new()
# add a set
i <- Set$new(m, "i")
# access records
i_recs <- i$records
}
