% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Set.R
\name{Set}
\alias{Set}
\title{Set Class}
\description{
A class for Set objects. This class inherits from an abstract Symbol class.
The documentation for methods common to all symbols can be accessed via help(.Symbol).
Please visit https://www.gams.com/latest/docs/API_R_GAMSTRANSFER.html 
for detailed documentation of this package.
}
\examples{
# create a container
m = Container$new()
# add a set
i = Set$new(m, "i")
# access records
i$records
}
