% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UniverseAlias.R
\name{UniverseAlias}
\alias{UniverseAlias}
\title{UniverseAlias Class}
\description{
A class for Alias objects that are aliased to the Universe set.
Please visit https://transfer-r.readthedocs.io/en/latest/
for detailed documentation of this package.

#' @examples
# create a container
m = Container$new()
# add a UniverseAlias
u = UniverseAlias$new(m, "u")
}
