% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtr_R6.R
\name{Variable}
\alias{Variable}
\title{Variable Class}
\description{
A class for Variable objects. This class inherits from an abstract 
symbol class.To access the documentation for any of the following methods, 
use help(Symbol).
countEPS, countNA, countNeginf, countPosinf, countUndef, getCardinality,
getSparsity, getMaxValue, getMinValue, getMeanValue, getMaxAbsValue,
isValid, toDense, whereMax, whereMaxAbs, whereMin.
}
\section{Super class}{
\code{\link[gamstransfer:Symbol]{gamstransfer::Symbol}} -> \code{Variable}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{description}}{description of symbol}

\item{\code{dimension}}{of symbol}

\item{\code{domainForwarding}}{flag that forces set elements to be recursively 
included in all parent sets (i.e., implicit set growth)}

\item{\code{domainLabels}}{column headings for the records dataframe}

\item{\code{domainNames}}{string version of domain names}

\item{\code{domainType}}{none, relaxed or regular depending on state of domain links}

\item{\code{name}}{name of symbol}

\item{\code{numberRecords}}{number of symbol records}

\item{\code{records}}{the main symbol records}

\item{\code{refContainer}}{reference to the Container that the symbol belongs to}

\item{\code{shape}}{a list describing the array dimensions if records were
converted with $toDense()}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{description}}{description of symbol}

\item{\code{dimension}}{of symbol}

\item{\code{domainForwarding}}{flag that forces set elements to be recursively 
included in all parent sets (i.e., implicit set growth)}

\item{\code{domainLabels}}{column headings for the records dataframe}

\item{\code{domainNames}}{string version of domain names}

\item{\code{domainType}}{none, relaxed or regular depending on state of domain links}

\item{\code{name}}{name of symbol}

\item{\code{numberRecords}}{number of symbol records}

\item{\code{records}}{the main symbol records}

\item{\code{refContainer}}{reference to the Container that the symbol belongs to}

\item{\code{shape}}{a list describing the array dimensions if records were
converted with $toDense()}

\item{\code{summary}}{output a list of only the metadata}

\item{\code{type}}{type of variable (string)}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Variable$new()}}
\item \href{#method-setRecords}{\code{Variable$setRecords()}}
\item \href{#method-clone}{\code{Variable$clone()}}
}
}
\if{html}{
\out{<details ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="gamstransfer" data-topic="Symbol" data-id=".linkDomainCategories">}\href{../../gamstransfer/html/Symbol.html#method-.linkDomainCategories}{\code{gamstransfer::Symbol$.linkDomainCategories()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="gamstransfer" data-topic="Symbol" data-id="countEps">}\href{../../gamstransfer/html/Symbol.html#method-countEps}{\code{gamstransfer::Symbol$countEps()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="gamstransfer" data-topic="Symbol" data-id="countNA">}\href{../../gamstransfer/html/Symbol.html#method-countNA}{\code{gamstransfer::Symbol$countNA()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="gamstransfer" data-topic="Symbol" data-id="countNeginf">}\href{../../gamstransfer/html/Symbol.html#method-countNeginf}{\code{gamstransfer::Symbol$countNeginf()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="gamstransfer" data-topic="Symbol" data-id="countPosinf">}\href{../../gamstransfer/html/Symbol.html#method-countPosinf}{\code{gamstransfer::Symbol$countPosinf()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="gamstransfer" data-topic="Symbol" data-id="countUndef">}\href{../../gamstransfer/html/Symbol.html#method-countUndef}{\code{gamstransfer::Symbol$countUndef()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="gamstransfer" data-topic="Symbol" data-id="findDomainViolations">}\href{../../gamstransfer/html/Symbol.html#method-findDomainViolations}{\code{gamstransfer::Symbol$findDomainViolations()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="gamstransfer" data-topic="Symbol" data-id="getCardinality">}\href{../../gamstransfer/html/Symbol.html#method-getCardinality}{\code{gamstransfer::Symbol$getCardinality()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="gamstransfer" data-topic="Symbol" data-id="getMaxAbsValue">}\href{../../gamstransfer/html/Symbol.html#method-getMaxAbsValue}{\code{gamstransfer::Symbol$getMaxAbsValue()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="gamstransfer" data-topic="Symbol" data-id="getMaxValue">}\href{../../gamstransfer/html/Symbol.html#method-getMaxValue}{\code{gamstransfer::Symbol$getMaxValue()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="gamstransfer" data-topic="Symbol" data-id="getMeanValue">}\href{../../gamstransfer/html/Symbol.html#method-getMeanValue}{\code{gamstransfer::Symbol$getMeanValue()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="gamstransfer" data-topic="Symbol" data-id="getMinValue">}\href{../../gamstransfer/html/Symbol.html#method-getMinValue}{\code{gamstransfer::Symbol$getMinValue()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="gamstransfer" data-topic="Symbol" data-id="getSparsity">}\href{../../gamstransfer/html/Symbol.html#method-getSparsity}{\code{gamstransfer::Symbol$getSparsity()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="gamstransfer" data-topic="Symbol" data-id="isValid">}\href{../../gamstransfer/html/Symbol.html#method-isValid}{\code{gamstransfer::Symbol$isValid()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="gamstransfer" data-topic="Symbol" data-id="shape">}\href{../../gamstransfer/html/Symbol.html#method-shape}{\code{gamstransfer::Symbol$shape()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="gamstransfer" data-topic="Symbol" data-id="toDense">}\href{../../gamstransfer/html/Symbol.html#method-toDense}{\code{gamstransfer::Symbol$toDense()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="gamstransfer" data-topic="Symbol" data-id="whereMax">}\href{../../gamstransfer/html/Symbol.html#method-whereMax}{\code{gamstransfer::Symbol$whereMax()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="gamstransfer" data-topic="Symbol" data-id="whereMaxAbs">}\href{../../gamstransfer/html/Symbol.html#method-whereMaxAbs}{\code{gamstransfer::Symbol$whereMaxAbs()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="gamstransfer" data-topic="Symbol" data-id="whereMin">}\href{../../gamstransfer/html/Symbol.html#method-whereMin}{\code{gamstransfer::Symbol$whereMin()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
There are two different ways to create a GAMS Variable and 
add it to a Container. One is using the Variable constructor and 
the other is using addVariable method which calls the Parameter 
constructor internally.
addVariable is a Container method to add a Variable.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Variable$new(
  container = NULL,
  name = NULL,
  type = "free",
  domain = list(),
  records = NULL,
  domainForwarding = FALSE,
  description = ""
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{container}}{A reference to the Container object that the 
symbol is being added to}

\item{\code{name}}{string argument for name of the Variable}

\item{\code{type}}{Type of variable being created [binary, integer, 
positive, negative, free, sos1, sos2, semicont, semiint]. The default
is "free"}

\item{\code{domain}}{an optional argument specifying a list of strings, 
a string. default value is "*".}

\item{\code{records}}{specify set records as a string vector or a dataframe.}

\item{\code{domainForwarding}}{an optional logical argument to specify 
domain forwarding. Default value is FALSE.}

\item{\code{description}}{string specifying description for the set}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
a Variable object
main convenience method to set standard dataframe formatted records
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-setRecords"></a>}}
\if{latex}{\out{\hypertarget{method-setRecords}{}}}
\subsection{Method \code{setRecords()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Variable$setRecords(records)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{records}}{specify set records as a vector, matrix, 
array or a dataframe.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Variable$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
